Move letter renderer into Letter
